Drop .pisi suffix, use .eopkg suffix
This is to save confusion with other distributions that may have in
the past used PiSi, or indeed still do. Removing compatibility with
these foreign .pisi files ensures there is no room for breakage
within Evolve OS, and does not tie Evolve OS to some obligation
of compatibility for a standard that only Evolve OS is now maintaing,
and SolusOS maintained previously.

This is the first step in the road away from the legacy format and
concepts.