Export pkg-config information when building packages
"PkgConfig" fields may be manually added to the Provides of each package
to indicate which pgkconfig files they export, however it is more suitable
to leave this task to PiSi. When building a package that installs a
.pc file, the relevant information is already exported and provided
in the eopkg-index.

This will be followed by changes enabling developers to use these new
pkgconfig exports as build dependencies, greatly reducing package
complexity.