build: Automatically scan and add binary dependencies
PiSi will now scan LSB executables and shared libraries during packaging
to ensure all binary dependencies are accounted for. Note this may add
a minute or so onto larger builds, but it really is worth it for the
improved dependencies.

Note that dependencies are resolved against currently installed packages,
which is the correct behaviour. The assumption is that you link against
installed packages, and if something is still missing you should really
check you have a clean build environment..