build: Try both lib and lib64, also skip "not found" deps
Under Evolve OS, for some unknown reason the "not found" string with
random dependencies will cause gobject-introspection to become a dependency
of the package.

This change ensures there is a linked library to search for now, as well
as ensuring that "/lib/" and "/lib64/" links both work for the same
packages. Eventually this compatibility will disappear when all packages
within Evolve OS have been forced into /usr/lib/