build: Refresh installdb for "unaccounted" pkgconfig files
Currently there is an issue where an entirely fresh evbuilder session
will not add pkg-config dependencies due to "unaccounted" pkgconfig
files, as the installdb is out of date due to newly installed
build dependencies.

We now explicitly refresh the installdb to ensure we're dealing with
the correct set of dependencies.