build: Ensure we set automatic dependency minimum release number
Given that our automatic dependencies are either binary or pkgconfig, we
need to ensure that upgrades are done in a sensible fashion. For binary
dependencies we'll now set the releaseFrom to be that of the release number
of the package we built against.

This greatly helps in mass-updates that are co-dependent, ensuring that
when required, an entire "section" (packages and their dependencies) are
updated to the correct release.