db/packagedb: Drastically improve performance of pkgconfig search
We now provide a far simpler approach to pkgconfig searching, by
checking the repodb doc tags for Providers. This is a much smaller
database to work with than the individual package providers, so
yields far better numbers.

The existing methods are ported lazily to the new API, to ensure that
nothing breaks during this transition. However, all consumers of the
API (ypkg, etc) should immediately move to this new API to decrease
build times.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>