db/filesdb: Add new replacement methods for finding .pc providers
Solus follows standard rules to determine where a pkgconfig provider
is defined, by having a .pc suffix in certain pkgconfig directories.
We now make use of these standard names by inquiring the filesdb
directly to see if it contains one of these files.

This new API should be used instead of the packagedb API which is
very, very slow. For a simple lookup on a single record, which is
known to not exist, it would take 4.52s on a Haswell machine. It
now takes under 0.2s to initialise the database & Python, and to
query all pkgconfig providers in the packagedb, before checking
the filesdb to determine if it exists or not.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>