Remove interaction with `COMAR` component
This removes eopkg support for interacting directly with COMAR, and
instead will only allow COMAR to talk to pisi/eopkg. Instead, all
configuration requests will be proxied to `usysconf` at the end of
each transaction.

We achieve this by wrapping the currently unsafe transactions in
a try/finally block and make sure usysconf will ALWAYS run for any
actionable transaction. After that, it's entirely up to usysconf to
do something with the system.

Note that due to the way pisi has traditionally used COMAR, we cannot
realistically delete the comariface file. This is imported dynamically
at runtime by the existing eopkg on disk, and failure to import it
with the old process would break the update, resulting in comar-api
issues.

As such we safe-guard that transition element of the process, and ensure
the newly run eopkg will never talk to COMAR for postinstall tasks again.
This will require some jostling from kernel-glue package to unconditionally
run `usysconf` during the very last use of COMAR to ensure system
consistency until the new eopkg takes over.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>