installdb: Flag any package that is automatically installed
This will let us add flags for each package that was somehow automatically
installed on the system, i.e. as an automatic dependency of another Thing
that wasn't within the set of packages requested for installation.

This will be useful in future when we add a command to analyze the system
for orphan packages, i.e. a plan remove for all the orphan names that can
be removed safely because they don't have reverse dependencies outside of
the flagged set.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>