Fix issue where permission error is marked as broken
If a path is inside a directory that we don't have permission to read,
eopkg marks the related package as "Broken" instead of "Unknown".

This ensures that packages with files that we cannot read are always
marked as "Unknown".

This namely fixes an issue where gnome-control-center, gvfs, libvirt,
udisks, flatpak, and possibly other packages could be marked as
"Broken" when running `eopkg check` without root.