Backported patch from dev.getsol.us to allow fix file replacing with folders.
In some rare instances, a file will be replaced with an identically named folder. This commit allows eopkg to semi-gracefully handle this by catching the IOError exception that is raised by the 'tarfile' library and carrying on as if it were a normal OSError for an already existing file.

Commit thanks to @DataDrake on 20.08.2018.