fetcher: add retry mechanism and reduce timeout to 10s to workaround RIT issue
urllib2 doesn't provide a nice retry mechanism for us unlike urllib3 so we have to
create our own in an ugly while loop. It's a bit blurgh but does the job.

For the RIT issue we actually have to catch SSL errors specifically rather than
generic urllib2 or socket timeout errors. We could potentially expand this retry
mechanism to those errors as well.

Be aware this is currently definitely a workaround for the current RIT woes.