Add -fno-plt to default build flags
As we default to BIND_NOW and symbolic-functions we already skip
the procedural lookup table already, the PLT mostly sits useless,
removing it will marginally decrease package sizes and improve start
up times.

A few packages that employ lazy-linking (xorg) or programs doing
funky things (e.g. valgrind) will need the PLT so we'll add an opt-in
to ypkg for those packages.