Also install updates to revdeps of rundeps of the package(s) to install
If updates are available, go through installed revdeps of the rundeps for
the package(s) to install and add them to the install graph. PiSi assumes
that the latest versions of the rundeps are suitable for installation
without considering the revdeps of deps for packages already installed.
Previously, a user always had to do a full update first to guarantee a
stable system.

Example: package A and packages C, D and E depending on A are available
for update. User installs package B depending on A and A will be updated.
However, other installed packages C, D, E depending on A will not be.
In the case of so name rebuilds this can lead to a very broken system
and is fairly fundamental flaw in PiSi.

Doing this is obviously slower, taking between 2-3 seconds for most
packages as opposed to 0.6-1.5secs to calculate the deps if no updates
are available, with the worst case tested around 11secs.

With this patch on a system with ~380 updates available, it could
successfully install packages without updating first. Previously,
doing so would break the install.