Ensure the install order is correct and optimize
Follow up to commit bc8b9ab.

Previously, the rev deps of rundeps check would add all needed
dependencies to the install graph but had the chance to install
them in the wrong order. E.g. if the package to install A, depends
on B and C, and B also depends on C. There was a possibility that
B would be installed before C.

To resolve this we append dep to checked after it has gone through
the rev deps of the rundeps to ensure all revdeps are met.

This opens up an optimization opportunity where we only check
revdeps if dep is in the list of available updates and hasn't been
checked already.

These changes make the general case a lot faster and the worst case
a few seconds slower. They also has the benefit of not downgrading
package updates a packager or user may have installed locally for
testing but do not yet exist in the repo.