ops/install: Ugly workaround for file conflicts edge-case
Turns out the upgrade op simply ignores file conflicts assuming they're
vetted. Now that install also upgrades revdeps of deps we can hit cases
where we hit file conflicts. In the case of a file moving from pkg A
to pkg B, we want the operation to continue. However, pisi has no way of
knowing the list of files for non-installed pkgs. Additionally, we do not
want to simply want to ignore file conflicts completely for the install op
like the upgrade op does currently.

To "resolve" this edge-case if the conflicting pkg is in the list of
upgradable pkgs then we'll assuming the file conflict is not an issue,
and it has been handled by the packager.

The issue:
- pkg B v2 is attempting to install /usr/bin/foo but /usr/bin/foo already exists on
  the system from pkg A v1
- pkg A v2 also is scheduled to install and no longer contains /usr/bin/foo but pkg B
  v2 is scheduled ahead of it.

To resolve properly:
- Extract pkg A v2 (it should already be downloaded)
- Check if /usr/bin/foo still exists by comparing files of pkg A v2 to filesdb pkg A v1.
- If it doesn't, allow operation to continue, else abort.