pisi/autoxml.py: Restore decoding with a dummy node
Due to a misunderstanding from my side, in a previous commit I thought
decoding a node using a dummy parent was silly. While this is true,
the code is structured in such a way this is necessary, and bending
the code to avoid that is so hard to the point it's not worth it.