Add a bunch of prints showing the license issue
From what I can tell, the issue is that pisi/pxml/xmlext.py/addNode is
buggy; instead of always appending a _new_ childe element to the parent
node, it only adds a new element to the parent node on the first run.

On subsequent runs, it keeps overwriting the first (and only) existing
child node it seems.

Signed-off-by: Rune Morling <ermo@serpentos.com>