archive.py: Re-encode file_name object to utf-8
This is done at the last possible moment before interfacing with the py3
standard library tarfile.py module to keep pisi py3 changes to a minimum.

Lucky for us, it was enough to simply re-encode the decoded-to-latin-1
file_name string to utf-8, which the tarfile.py module can then
successfully handle.

The bug was exposed by the usdx package.yml, which includes filenames
with umlauts in them that tarfile.py would choke on prior to this commit.

Test Plan:

1) clone/update the ypkg.git repo
2) run `./prepare_ypkg_test_venv.sh`
3) check that the venv is active!
4) git -C ../eopkg/ checkout python3-utf8-filename-decoding
5) Install needed usdx deps locally with something like:
   `sudo ./ypkg-install-deps ../soluspkgs/packages/u/usdx/package.yml`
6) Test the usdx package build with something like:
   `fakeroot ./ypkg-build ../soluspkgs/packages/u/usdx/package.yml`
7) If usdx builds as expected, remember to do a:
   `sudo eopkg history -t <usdx builddeps install transaction above)>

Signed-off-by: Rune Morling <ermo@serpentos.com>