venv: Tweak build_chroot logic and help messages
We now explicitly point out that the ./start_chroot.sh script exists.

In ./start_chroot.sh, we now run 'usysconf run -f' prior to dropping the
user to a root login shell in the chroot and we ensure that /run is
mounted in the chroot as well.

The above togetther ensures that eopkg4-bin now functions out of the box
in the chroot in my local tests.

Signed-off-by: Rune Morling <ermo@serpentos.com>