chroot scripts: Refactor and mount/unmount tweaks
Split out shared functions to shared_functions.bash for convenience and
source it from scripts that need it.

Add `trap <unmount function> EXIT` to the chroot scripts so CTRL+C
doesn't leave the system in a weird state with partially mounted virtual
file systems.

Finally, be very careful and deliberate about the mount and unmount order
of the virtual filesystem.

Signed-off-by: Rune Morling <ermo@serpentos.com>