build_chroot: Focus exclusively on eopkg.py3 tests
This implies that extra commands can be added that focus on testing the
venv version of eopkg.py3, including deliberately adding a battery of
remove, history, etc. etc. commands to the build_chroot.sh script and
then chrooting in via systemd-nspawn to check that the root still works.

We could even generate a list of in-root commands to run in a shell
script in root/ in the chroot to verify stuff if we wanted to.

Signed-off-by: Rune Morling <ermo@serpentos.com>