Force install/upgrade order for eopkg & baselayout
This will ensure that future upgrades to eopkg and baselayout will always
happen in a deterministic fashion.

If eopkg is part of the install or upgrade transaction, eopkg will be
moved to the front of the transaction.

If baselayout is part of the install or upgrade transaction, baselayout
will be moved to the front of the transaction (possibly in front of
eopkg).

Note that testing this commit requires you to do a system wide pip
install until we get the ordered-set package included in the repo:

    sudo pip install ordered-set

Signed-off-by: Rune Morling <ermo@serpentos.com>