Py3: Deterministic install/upgrade order
This will ensure that future upgrades to eopkg and baselayout will always
happen in a (semi-)deterministic fashion.

- If eopkg is part of the install or upgrade transaction, eopkg will be
  moved to the front of the transaction.
- If baselayout is part of the install or upgrade transaction, baselayout
  will be moved to the front of the transaction (possibly in front of
  eopkg).

For context, only the ordering between eopkg, baselayout and any
system.base packages which are part of the transaction is currently
preserved.

The install order in the presence of system.base packages can potentially
also be tweaked now that we're using ordered sets where it matters (if
people are not happy with the current logic).

Note that this commit requires you to install the python-ordered-sets
package.

Signed-off-by: Rune Morling <ermo@serpentos.com>