filesdb: Default to FILESDB_PICKLE_PROTOCOL_VERSION=0
This is to make it more obvious what pickle protocol version a given
pisi FilesDB `files.db` database file will be written with.

Note that this does not _change_ the existing default.

Rather, it just makes the existing implicit pickle protocol version
(set per the python2 defaults) explicit in the code.

Signed-off-by: Rune Morling <ermo@serpentos.com>