FilesDB: Refactor for better encapsulation
With this change, the FilesDB ctor accepts a force_rebuild boolean, which
controls whether to force-rebuild the gdbm format on-disk db file.

The FilesDB ctor has been rebuilt to use a "collector pattern", based on
an analysis of valid states. It should hopefully make the logic of when
to rebuild (and when not!) a bit easier to follow.

As an added bonus, `pisi.api.rebuild_db()` is way less ugly now.

This commit also makes `sudo ./eopkg.py2 -dy rdb` work again on my system.

Signed-off-by: Rune Morling <ermo@serpentos.com>