FilesDB: Refactor for better encapsulation, output
With this change, the FilesDB ctor accepts a force_rebuild boolean, which
controls whether to force-rebuild the gdbm format on-disk db file.

The FilesDB ctor has been rebuilt to use a "collector pattern", based on
an analysis of valid states. It should hopefully make the logic of when
to rebuild (and when not!) a bit easier to follow, not to mention that it
should now output more useful informational messages re. the state of the
FilesDB on disk.

As an added bonus, `pisi.api.rebuild_db()` is way less ugly now.

This commit also makes `sudo ./eopkg.py3 -dy rdb` output somewhat more
useful in both normal and debugging mode.

Signed-off-by: Rune Morling <ermo@serpentos.com>