api,cli: Add checking for active repositories
Some commands require an active repository to be configured on the
system in order to be able to function. If no repositories are
configured, they may fail silently, or in other unexpected ways. This
solves that issue by adding an API function to check for the presence of
active repositories, and having the CLI bail with a message if none are
found.

This may not encompass the entire list of commands that need to check
for repositories, but at least it's a good start.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>