signalhandler: Refactor signal handling
This refactors the Pisi signal handler to be easier to follow. I'm not
actually sure it was ever working correctly before. I'm not sure if we
really need or want to hold pending signals to emit later, but I kept
that functionality, while providing a way to disable it.

As a part of this, we now disable keyboard interrupts (SIGINT) for more
operations: rebuild-database, remove package, and update repos.

Fixes https://github.com/getsolus/eopkg/issues/141
Fixes https://github.com/getsolus/eopkg/issues/105

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>