db/packagedb: When getting pkgconfig providers reverse through repo list
To ensure repo priority is respected.

Currently, repo priority is basically implied throughout the pisi
codebase unfortunately. However, trawling through the repo list in
reverse we ensure that we return the correct pkgconfig provider for the
highest priority repo.

E.g. repo A and B provide pkgconfig(foo), in repo A pkgconfig(foo) is
provided by foo1 and in repo B pkgconfig(foo) is provided by foo2. If
repo B has a higher repo priority than repo B then we want to ensure we
return the package foo2. As it's a dict where there can only be one
unique key. By trawling in reverse we override it by getting the
pkgconfig from the higher priority repo.

This fixes issues in packaging with local repos where pkgconfigs move
between packages.

Resolves #153.

This PR also removes a useless repodb redeclaration that was introduced
in the previous PR.