dist/completions/bash: Optimize getting packages for operations
Previously we parsed the xml file for _every_ tab completion event,
in no short words this was stupid and slow.

Instead, we populate the _eopkg_{i,p}db_list variable if it's not set
and populate it globally instead of locally so it can be re-used

Additional Changes:
- No longer return an empty package list if there are .eopkgs in current dir

The one downside of this change is there is no way to invalidate the current
list if the {package,install}DBs have changed. To recify this we could store
the mtime of the eopkg xml index files and clear the _eopkg_{i,p}db_list variables
in that case.